module_data_acquisition_websocket/requirement.txt
python-binance # 幣安 API 庫
google-cloud-pubsub # Google Cloud Pub/Sub 客戶端庫
google-cloud-secret-manager # Google Cloud Secret Manager 客戶端庫
